+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+# Release 1.26.0 29-10-2018
+
+- Correctly convert gRPC custom Status to JSON in transcoding. (#479)
+
+- Add start_esp flag: experiental_proxy_backend_host_header. (#476)
+
+- Add an option to enable HSTS (HTTP Strict Transport Security) (#474)
+
+- Add enable_debug flag (#471)
+
+- Not to log access token (#470)
+
+- Upgrade gRPC from 1.4.2 to 1.15.0 (#450)
+
+- CORS always adds the Access-Control-Allow-Origin header (#449)
+
 # Release 1.25.0 17-10-2018
 
 - Change side-way http to use 1.1 (#453)